refactor

+++ src/main/java/com/moviesandchill/portalbackendservice/dto/film/film/FilmDto.java
@@ -34,6 +34,6 @@ public class FilmDto {
 
     private String description;
 
-    private Double rating;
+    private Float rating;
 
 }
